Use pixel values for global header overrides
This fixes a sizing issue with the new Rizzo Next global header in the shop. The issue is that the sizing is too big due to the use of `rem` values in Rizzo Next.

Rizzo Next uses a base font size of `10px` for rems, Rizzo uses `.875em` and the shop uses `100%`. Removing the scaling factor ratio and using `px` for values fixes the sizing problem on the shop. We don't want to alter the base font size of the shop because that could have unintended side effects.

In case this issue were to happen somewhere else, using `px` provides consistent rendering across all apps.