Import Rizzo Next components into core_legacy
Rizzo Next icon fonts are not loading for the shop.

The shop appears to be using a different version of `core.css` and `core.css` is where the icon font is imported.

- Shop: `https://secure.lonelyplanet.com/static-ui/style/core.css`
- Home page: `http://assets.staticlp.com/assets/core-1a9d0baf1ebf141e8093a8e15fdafd9e.css`

However, shop has a `core_legacy.css` which is served from `assets.staticlp.com`:

- `http://assets.staticlp.com/assets/core_legacy-114d2aa6c48b6c84c847a250a5e2a125.css`

So, the hope is that importing the icon fonts into `core_legacy.css` will fix the issue.