Import Rizzo Next into app_core_legacy.js
Similar to the issue with the Rizzo Next icon fonts and the shop, it appears that the shop is using `app_core_legacy.js` instead of `app_core.js` like other apps. `app_core.js` is where Rizzo Next JS is being loaded, so adding it to `app_core_legacy.js` should add the JS to the shop.